+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @swisspost/design-system-styles
 
+## 10.0.0-next.60
+
+### Major Changes
+
+- Fixed Safari-specific issue where date/time input placeholders appeared in yellow instead of inheriting the correct color. (by [@alionazherdetska](https://github.com/alionazherdetska) with [#6974](https://github.com/swisspost/design-system/pull/6974))
+
+### Patch Changes
+
+- Aligned the local header icons padding with design specs. (by [@leagrdv](https://github.com/leagrdv) with [#7073](https://github.com/swisspost/design-system/pull/7073))
+- Updated dependencies:
+  - @swisspost/design-system-icons@10.0.0-next.60
+
 ## 10.0.0-next.59
 
 ### Patch Changes